% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfactorial.R
\name{myfactorial}
\alias{myfactorial}
\title{Calculate the Factorial of n by Recursion}
\usage{
myfactorial(n)
}
\arguments{
\item{n}{the number n which needs to calculate its factorial}
}
\value{
the factorial of the number n
}
\description{
Calculate the Factorial of n by Recursion
}
\examples{
myfactorial(5)
}
